get rid of catalog singleton
